Make doubly sure that a pointer is not dereferenced if invalid.
I know what you are going to say. But there is a Q_ASSERT right
there. Fine. But when Amarok is compiled w/out debugging that
assert will melt to only a debug printout. And Amarok will
crash. Besides the chance of a bad pointer here is really low.
So why not add suspenders to the Q_ASSERT belt.

svn path=/trunk/extragear/multimedia/amarok/; revision=872245