Add one more to size of dynamically allocated strings.
Mp3tunesLockerMeta uses strlen to determine the length
of the string that is copying. But strlen leaves off the
null terminating character. Which means that the new object
will have strings that are just one char too short. This commit
adds one more to the string length so it is properly sized.

svn path=/trunk/extragear/multimedia/amarok/; revision=872247