Restructure commented out code to make sure unreachable code is not compiled.
If code is unreachable, it should be commented out so that it is not
compiled into the binary to avoid bloat. Here are a few instances
where that is the case. The code was obviously made unreachable
on purpose. So it only makes since to completely comment it out.
Consider removing it or reworking it eventually.

svn path=/trunk/extragear/multimedia/amarok/; revision=872248