Revert changes to GroupingProxy made in r872993
Equivalency of album title is not the same as equivalency of album.  I
have six cases in my collection where two different artists have albums
with the same title, and Wikipedia lists hundreds of examples more.  And
we can't just test Album and Artist because this will screw up Various
Artist albums.  If testing equivalency of AlbumPtr in the GroupingProxy
fails to match tracks from the same album, then the fix needs to happen
in whatever code is creating the AlbumPtr and associating it with the
track.

CCMAIL: nhnfreespirit@gmail.com

svn path=/trunk/extragear/multimedia/amarok/; revision=873054