Remove unneeded const.
When there is a const at the end of a function declaration. For example:
QString foo() const {}

Then a const at the beginning is unneeded:
const QString foo() const {}

There were a few examples of this in the three files in the commit. All
have had one of their const removed.

svn path=/trunk/extragear/multimedia/amarok/; revision=874243