Unwind circular include directives.
It is possible in C/C++ to have two header files include
each other. Typically this is not a problem. The include
guards keep include files from being placed w/in the source code
twice and they generally do not get in each other's way. But
it is possible to muck it up and with the power of forward
declaration actually change the way functions come out.

To avoid this we can unwind any circular include directives.
This commit does just that. But to get the tree to compile
again some includes have to be added. Sometimes declarations
have to be used. And this commit does that too. Otherwise there
is no substantive change to the tree.

svn path=/trunk/extragear/multimedia/amarok/; revision=875907