new last.fm code for the last.fm service. basically, this replaces our old last.fm code drop from the 1.4 client with the  new modular liblastfm from max at last.fm. currently the library is included here in-source, and we are compiling it as a lib and linking to it internally. as soon as they can, last.fm will release it as liblastfm, and amarok will depend on it. for now, and 2.0, this will stay in-source. note that i include a FindLibLastFm.cmake that is not being used, as soon as we switch to the external lib we can just switch to searching for the external lib w/ no hassle.
NOTE: *please* test. i've tried to stress it as much as i can, but i can't think of all the edge cases. break it!

this should fix a number of bugs with the old client. there is 1 bug that i know of with this code, and i have email max.

BUG: 164156
BUG: 157296
BUG: 155197
BUG: 154904
BUG: 170816

svn path=/trunk/extragear/multimedia/amarok/; revision=877693