Made the settings dialog remember its current page.
The dialog will now remember which page was last used, and will show the same page again the next time you open it. However, it will not remember this across sessions, by design. After startup it will always show the first page (General Config), except on the first run, where it will show the Collection Config.

This is the same behaviour as in Amarok 1, and was discussed with our usability guru.

Additionally I've refactored some of the code and made it more modern and readable.

svn path=/trunk/extragear/multimedia/amarok/; revision=879452