Made the service browser show the services alphabetically sorted. Added a custom QSortFilterProxyModel implementation to achieve this.
NOTE:
There is a known regression being introduced by this commit: For every service item one extra item is displayed, which is empty. nhnFreespirit and I have agreed on committing this to trunk anyway, to get some more eyeballs on the problem. If we cannot sort the problem out, we are going to revert this commit again.

svn path=/trunk/extragear/multimedia/amarok/; revision=881494