m_driver should be initialized in the ctor
The ctor for FastForwardWorker neglects to initialize m_driver. Since it
is a enum type, it should be initialized to something. The SQLite was
selected since it is the typical type for Amarok 1.4. The other QStrings
were initialized to QString::null (by explicitly executing their ctor)
just to make extra sure.

svn path=/trunk/extragear/multimedia/amarok/; revision=887357