Revert "Update containments geometries before changing to zoom out level." Revert "Remove old hack not needed anymore." Revert "Fix nasty bug in CV that showed misplaced containments after resizing and changing containments."
This last-minute bugfix introduced regressions. As it is a large code drop literally HOURS before tagging 2.0 final, this is absolutely the wrong time to do this and so must be reverted. It randomly breaks the size of the containments so that applets are not always added in the right place, or restored correctly on startup, or other weirdness. This needs more testing before committing.

Liw-, i understand that this fixes the browser-collapse bug, but we can't deal with a mess of bugs for 2.0 final. Post-2.0, please recommit and we will work out the weirdness, but for now, this has gotta go.

svn path=/trunk/extragear/multimedia/amarok/; revision=892923