Fixed problems with the collection scanner missing part of the collection, or potentially failing altogether.
Specifically the problem was with restarting the amarokcollectionscanner
binary after it has crashed, which we are doing transparently in the
background, without the user ever noticing (normally). The scanner can
crash due to TagLib crashing, which still can happen sometimes.

The code for restarting the scanner was never properly ported over from
Amarok 1, and it would often miss a part of the data during restarting.
This should be fixed now.

PS: This took me about 4 days of intense hacking, so I'm really happy
now that it's done.

Please test :)

BUG: 176154

svn path=/trunk/extragear/multimedia/amarok/; revision=903687