Revert "Attempt to fix CV bug that showed misplaced containments after resizing and changing containments. "
This patch introduced serious regressions with the vertical sizing of
applets. This is the second time that I'm reverting this patch.

William, to be honest with you, this wasn't very nice of you to commit the same
buggy patch twice, and then letting others try to fix it up. I spent
about two hours on that today. Time that I could have invested better.

CCBUG: 176101
CCMAIL: vianasw@gmail.com
CCMAIL: amarok-devel@kde.org

svn path=/trunk/extragear/multimedia/amarok/; revision=904193