Bug fix: If the playlist ends, make sure that the progress-slider is reset.
The cause of this bug was a regression from the following commit:
SVN commit 899127:
"Fix progress slider not showing when skipping last.fm streams. This may
or may not affect a number of other progress bar related bugs, please
check."

But the real root of the problem was that we didn't notify
EngineObservers with playbackEnded() when a track ended. So I changed
this now. Please watch out for regressions and inform me if you should find
such.

I have tested different scenarios with the new code (e.g. Last.fm
radio), and so far it seems fine.

svn path=/trunk/extragear/multimedia/amarok/; revision=910442