ReplayGain FTW!
Make replay gain support actually do something by
(a) getting the data we stored out of the collection database
(b) using it when the track changes

Also, improve the storage of replay gain tags by storing NULL when they weren't present on the original track metadata.  This allows us to substitute the track gain for the album gain when the latter is requested but doesn't exist.


svn path=/trunk/extragear/multimedia/amarok/; revision=911684