Increase the width of the volume slider widget by a factor of 2x, as discussed with Leinir and nhnFreespirit.
Reasoning: For use with touch screens, the precision of the shorter
volume slider was insufficient. E.g. it was impossible to select 1%;
instead it jumped directly from 0% to 4%.

It would have been possible to increase the length as a special
customization for touch screens, but we discussed on IRC that this
change might also be beneficial for Amarok proper, so I'm committing it
directly to trunk.

svn path=/trunk/extragear/multimedia/amarok/; revision=919526