Fix tracks being counted twice, and track playback in the engine controller.
The enginecontroller is a mess.  This commit fixes things for me (which probably breaks things for at least one other person...)  Remove slotPrefinishMarkReached (which only called slotAboutToFinish) as it resulted in duplicate stat calculation.  Also, Phonon::MediaSource can have remote media enqueued, and in fact needs to to make playback work for http/lastfm streams. Please give this some testing and if it doesn't break anything I'll backport it to stable in a few days

BUG: 180205
CCMAIL: amarok-devel@kde.org

svn path=/trunk/extragear/multimedia/amarok/; revision=920478