Implement rename for groups.
This was very difficult since I turn a tree structure into a flat model and back to a tree model in the proxy.
I might turn SQL playlists into a flat structure buy removing the playlist_groups table and adding the column label to playlists, just like in SQL podcasts.

svn path=/trunk/extragear/multimedia/amarok/; revision=920820