NOTE: I *know* this messes up the PUD.  This will eventually be fixed.
Start to refactor the PUD a bit to make calculating geometry much easier.  Instead of using SVG sizes to calculate item sizes, use a svgElmenRect which can be 
programatically set and (eventually) scale the SVG to fit.  This simplifies a lot of logic.  However until I am done the PUD graphics are a bit borked (though 
they do work).

svn path=/trunk/extragear/multimedia/amarok/; revision=922407