Allow choosing between album and track replay gain modes, or even disabling replay gain altogether.
I put the option in the Settings menu, as that seemed the most appropriate place (we don't want to pollute Amarok, and it's not a Playlist option).

The other option is to put it in the configuration dialog, but users may want to change it when changing random mode between track and album, or when changing playlists.

There are no icons for the different modes in the menu yet.


svn path=/trunk/extragear/multimedia/amarok/; revision=924937