Improve the collectionscanner by a factor of a lot.
Treating QXmlStreamReader::NoError as a non-recoverable error is probably going to make things not work that well.  This was just one of those things that was way too easy a bug to find and fix for way too long, apparently.  I'm not certain this fixes all the issues, but it scans a lot more of my collection now.  Will backport to 2.0 branch after it gets some more testing.

CCBUG: 176154

svn path=/trunk/extragear/multimedia/amarok/; revision=928105