PLEASE TEST
Right now, directories that no longer exist are never removed from the database.  As far as I can tell this is a bug, not a feature (dynamic collection detects existing directories that are no longer mounted, and does not deal with deleted directories).  In fact, the code detecting directory deletion up to this point has done the exact same thing as the code that detects changed mtimes -- i.e., it tells the scanner to scan it (even though it doesn't exist).

This commit should fix that by deleting directories that no longer exist from the database during an incremental scan.  This works in my testing and intuitively it's correct, but please test this.

CCMAIL: amarok-devel@kde.org

svn path=/trunk/extragear/multimedia/amarok/; revision=935422