Continue work on a first-run-tutorial which was first started by Jeff Mitchell (jefferai).
We've currently reached the point where the tutorial displays one page
with nicely formatted text in it, and a Close button. Further pages (not
too many, I suggest) could be added, for explaining the basic usage of
Amarok (especially the 3 panes concept). Also graphics could be added
for WOW-factor.

I've deliberately left out i18n() for the text until we have it
finalized, so that this won't get too messy for translators.

To enable the tutorial, set "FirstRunTutorial" in amarokrc (General
section) = true.

CCMAIL: amarok-devel@kde.org

svn path=/trunk/extragear/multimedia/amarok/; revision=939066