Try to be smarter about Amarok not finding its plugins.
Now that we no longer use --noincremental for kbuildsycoca4, the call is
much faster, and also doesn't seem to fork. So instead of giving up and
showing the user an error, we simply run that, and then try again to
load our plugins :)

And if that still fails, we show the error dialog, as we used to do.

svn path=/trunk/extragear/multimedia/amarok/; revision=940393