Make double clicking a top level item in a  CollectionTreeView behave just like any other item. This once again makes it possible to add Cool Streams tracks (as this is a collection where the tracks _are_ the top level) as well as entire Magnatune genres and Ampache artist by double cliking.
This does behave a bit odly with top level items as these apparently have special code that expands/closes them on a single click. I suppose this was added to get rid of the "+" button on the new collection headers, but I am not 
quite sure I like this aproach. Perhaps having the "+" button is better after all?


svn path=/trunk/extragear/multimedia/amarok/; revision=943083