improved database debugging
this commit makes it possible to dump the contents of Amarok's database to CSV files
and dump tables related to the collection scanning before and after the scan.

the whole database can be dumped by calling Amarok.Collection.dumpDatabaseContent() in the script console.
Warning: it will try to delete a file in the home directory that has the same name as the CSV that it is about to write. Check DatabaseUpdater.cpp/SqlCollection.cpp for the file names that it will try to write.

Additionally, one can enable automatic dumping of relevant SqlCollection tables before and after a collection scan.
Add the following section to amarokrc:
[SqlCollection]
DebugDatabaseContent=true

This requires a restart

svn path=/trunk/extragear/multimedia/amarok/; revision=972963