The playlist dropped case is not handled right.
The playlist dropped case is not handled. Fine. But the code will actually
continue if the dynamic_cast didn't work. That means that the code will
segfault in the foreach loop. Not good. Better to change the code to do
what it should (check the dyn_cast for null return) and comment out the
whole if statement until a proper implementation can be put in.

svn path=/trunk/extragear/multimedia/amarok/; revision=978745