Don't crash when starting a collection scan after a previous scan had been aborted.
We destroyed the ScanManager instance when aborting the scan,
and then never instantiated it again, ending up with a 0-pointer dereference.
Now we reinstantiate the ScanManager automatically.

This seems the safest way to fix this issue for now. I pondered
refactoring the code to keep the ScanManager always alive, but this
would require major changes and seemed too risky at this point.

BUG: 195242

svn path=/trunk/extragear/multimedia/amarok/; revision=978819