Possibly fix all sorts of metadata bugs.
SqlRegistry acts as a cache that had severely broken cache validation.  So when results were scanned via the collection scanner, accesses of SqlTracks and the like from the registry could be stale, and could subsequently be used in UPDATE/REPLACE statements into the database (causing tracks to be "lost" or to be listed under the wrong artist) and even be used when writing metadata to tags (people complaining about "old tags being written").

There are a number of facets to this patch.  First, when a collection scan is run, all scanned tracks were already being kept for UID purposes; now, it takes this list and forces updates from the database for all of these tracks (which is sane since their fresh data has just been scanned).  Second, when track information is saved to the database, it now *only* writes the fields that have actually changed, instead of writing all fields stored in the track (hopefully the first part removes the need for this, but this is good for safety's sake, or in case the first part doesn't properly work).  Third, when writing metadata to a file, the cache that is used to hold items is cleared and only the item(s) being written are sent to the function for writing.

Finally: there is some extra debugging right now that got carried into this commit...I'll remove it later, but in the meantime if someone wants to use it to figure out why Observers only get updates to the URL of a track a single time (and then disappear off the Observer list), it'd be *swell*.

svn path=/trunk/extragear/multimedia/amarok/; revision=985896