Acoustic similarity subsystem
Adds support to Amarok for the GTACFeat[0] audio similarity library.
GTACFeat is not required, so if you don't have it, Amarok should
compile and function normally without it.

If you want to be able to use the acoustic similarity features, as of
right now you should install GTACFeat from the SVN sources; hopefully by
the time Amarok 2.2 is ready for release, a beta version of GTACFeat
will be available.  Also, all this code is just back-end stuff.  There
isn't anything yet available in the GUI that takes advantage of acoustic
similarity, though there will be in the very near future.

[0] http://sourceforge.net/projects/gtacfeat

CCMAIL: amarok@kde.org

svn path=/trunk/extragear/multimedia/amarok/; revision=987771