Revert Soren's patch.  (Note that this leaves the added files in-tree for now, just reverts the modified files...this is to avoid creating peg revisions unnecessarily).
Ian, Alejandro and myself (at least) have multiple reservations about the changes made in this patch.  It needs review and testing before being pushed out to trunk, especially as it updates the database (in a way 
that is probably not correct).  This really needs to be put in a git branch first and tested out, or submitted as a patch to the mailing list for review.

CCMAIL: amarok-devel@kde.org

svn path=/trunk/extragear/multimedia/amarok/; revision=987798