LicenseChecker: Added support for more than one valid license header. Also moved to supplementary_scripts.
This should pass any KDE e.v.-style license headers, as well as any others we choose to add in the future.

The initialisation of the QList<QStringList> used to hold the different valid license headers is quite ugly at the
moment. I couldn't work out how to initialise a QList with a list of values in one line.

svn path=/trunk/extragear/multimedia/amarok/; revision=987863