Insomnia leads to hacking.  Besides, did you really think I'd leave Gran Canaria without committing this?  :-D
If you dare (really), follow me down the rabbit hole of this bitch of a fix that's taken me a month to figure out...

Let's suppose that your file location changes, and you want things to be aware of this. So you set the new URL on the SqlTrack you fetched from the SqlRegistry using the uidUrl of the Track. This eventually causes observers to be notified. This ends up notifying the tracks displayed in the Collection Browser, although only if they're actually displayed, which means it's an intermittent problem (fun to track down). When these tracks get notified, for whatever reason, they don't pull the new information from the track but rather run a QueryMaker query to fetch the new updated track by re-querying the SQL database with the various artist/album/track information.  This pulls a URL from the database which is correct for the track, but which is then sent to SqlRegistry to get a cached copy of the track from the URL.  SqlRegistry says it doesn't have a cached copy, because it's a non-cache-coherent cache that apparently I didn't *really* make fully cache-coherent with my massive fixes in r985896 (it is now, promise, at least until I find the next place it's not coherent, at which point I will renege on that promise), and the map it looks in doesn't have that URL, it has the old one. So it creates a new track (possibly leaking the old one), returns that to the Collection Browser, and then that new track is put in the registry to be used in the future. And in the process, all of the Observers go bye-bye.

J.F.C.


svn path=/trunk/extragear/multimedia/amarok/; revision=992942