Added experimental "go back" button for browser widget.
Inspired by the Kickoff KDE menu, I've added a tall vertical button on
the left side of the browsers. Clicking it makes you go up one level in
the browser categories.

I think this makes it easier to navigate in the new browser, as the
breadcrumb widget alone is a bit cumbersome to use. However, making it
look nice is difficult (try uncommenting the style sheet line).

svn path=/trunk/extragear/multimedia/amarok/; revision=997690