Force on --as-needed.
Building from source tree is broken on some distros without --as-needed specified. Actually, it builds but it can't find collection plugins, which is bad. So force it on -- many distros are already doing this in
their packages.

Hopefully this won't cause issues for anyone -- if it does, we should see what the right resolution is (i.e. it might be fixing them, not reverting this, so don't revert blindly).