Clean up the EngineController logic.
This commit has been sitting on my computer for too long.

Leo: this does the fadeout logic the "right way" (and in doing so makes
it work properly on GStreamer), at the expense of not working on
CoreAudio (as I recall, there were some issues with keeping a
VolumeFaderEffect in the audio path on CoreAudio).  You might want to
try re-enabling it and seeing if Phonon on CoreAudio is still broken.

CCMAIL: lfranchi@kde.org