Suppress compiler warnings associated to casting
The test code does a lot of casting. Likely because the Qt compare
macros require the correct types. But the casting is unnecessary
when dealing w/ constants. C/C++ already has keywords to make sure
the constants are the right type (in this case U/u and F/f).

To belabor the point further, since this is C++ code using C like
casting '(type)variable' is probably not a good idea. Instead use
C++ like casting, 'type(variable)' or 'static_cast<type>(variable)'.