fc can be leaked in 'is infeasible' path
When the code checks to see if the filter is feasible, if it isn't
then the filter is not added to the list of feasible filters and
the function returns. This would leave fc undeleted and unmanaged.
Which would be a leak. When a filter is determined to be infeasible
it should be deleted.