t function parameter hidden by t declaration
t is the name of the Token function parameter. Then later in the code
t is declared as the Token variable for the foreach. This hides the
function parameter w/in the foreach scope. Which wouldn't be a problem
except that if someone later decides that they do want to use the
function parameter w/in that scope, they may accidentally not get what
they want. Changing the name to token for the foreach declaration makes
everyone happy!