m_scheme is heap allocated in the ctor
m_scheme is heap allocated in the ctor. That means that if it is not
to be leaked it needs to be deleted in the dtor. But SortScheme
does not have a declared dtor. So to do the delete, the dtor must
be declared and then defined w/ the delete included to do the job.