Make all cases in switch conform
In categoryIcon() the case of PodCast does not return the icon. Instead
it falls through and the dynamic icon is returned. If this is the
desired behavior it needs to be commented on. But I doubt it is the
desired behavior, so I stuck a return in front of the KICon ctor.