dynamic_cast should be checked.
As always, check dynamic_cast (in ProxyBase).

The AFTTagger.cpp change is based on the dynamic_cast. But it is
more of an efficiency change. By declaring pointers and assigning
them at the first dynamic_cast we can save the cpu resources of
not having to do the second dynamic_cast.