Change &= to |= in TimecodeTrack::set functions
It occurred to me tonight that the changes I had to do in TimecodeTrack
to correct the use of bitwise | to & were such that the writer had
confused the two operators and that there might be other such
mistakes in the same class in other functions. Sure enough, all of
the set functions used &=. Which would wipe out any other flags
in the field that had already been set. What I think the writer
really wanted was |= so that particular flag could be set in the
field. This commit changes that.