Only use inline once.
C++ really doesn't need the inline directive to be listed every time.
In fact listing it at declaration and definition is dangerous as it
can produce strange compiler warnings. In this case the setDbChanged
function had it listed twice and g++ was complaining about how the
function wasn't defined (when it definitely was).