daapreader: Remove conditionals that never evaluate false.
Because of the if clause at the top of the while loop, the
length ? blah : QString() foreach each of the string matches will always
evaluate true. This is effectively dead code. By removing it we
don't fool the developer into thinking it works and we create
smaller object files. Win win!