Don't send out dummy notifications: the receivers should be fixed instead
This completes the recommit of 28b3e64d1dc5c7170d7b5d09e6cb4c8b3cdd4e0 (in a more sensible split-up manner).  This commit was the most troublesome part of that commit, but I've been using it with the Xine engine for several days now with no trouble.