Filter out non-mediadevice tracks when saving a playlist to a mediadevice.
In 2.3 we should start a collectionlocation copy to the device for any non-MD track. But for now we are out of time and in string freeze.
The best sollution I could come up with is show the user the playlist is empty without any warning (a.k.a. non-translated string).

Possible prevents crash.

CCBUG:205951