Restore config dialog to acceptable size on KDE 4.3, hiding EQ settings.
As the equalizer only works with KDE 4.4 and its GUI takes
up too much space, and as we don't have the time or means (string
freeze) to fix it properly for 2.2.0, I reckoned the next best thing
is to hide the GUI completely on KDE 4.3. This should help the majority
of our users for now.

We're still showing the "Equalizer not supported" line, so it shouldn't
hurt discoverability.