Fix: Automatic playlist scrolling was broken in some cases.
The actual bug was very very nasty. EngineObserver used QSet for storing
observer pointers. This is what Qt docs say:
"QSet is unordered, so an iterator's sequence cannot be assumed to be predictable."
Now we're using QList, whose iteration order is predictable.

Thanks to Anthony Vital for pointing this out.

BUG: 193459