Fix Lyrics Applet saving function when lyrics is empty
When saving empty lyrics, make sure the lyrics cache is cleared for the
track. Previous behaviour is to bring back the last content (if there
was any) if the QTextEdit widget is emtpy.

If say, the wrong lyrics for a new track is fetched, and I went and delete
it in the editor, the next time the track is played the old lyrics would
come back up. This is unexpected.