Make "Toggle Main Window" work for minimised windows and multiple desktops
We want to show Amarok when it is hidden, and vice versa. There
are many window states to consider; when Amarok is:

    1. iconified (invisible)
    2. minimised (visible)
    3. on another desktop
    4. minimised and is on another desktop

For 1, the window is made visible (as before). For the others,
the window is unminimised and/or move to the current desktop if
applicable.

MainWindow::activate() and MainWindow::isReallyShown() are related
functions, the latter of which is not used at all within Amarok.

Qt doc for KDE::KWindowSystem says:

    Because of limitiations of the way Qt is implemented on Mac OSX, the
    WId's returned by methods in this class are not compatible with
    those expected by other Qt methods. So while it should be fine to
    pass WId's retrieved by for example calling the winId method on a
    QWidget to methods in this class the reverse is not true. You should
    never pass a WId obtained from this class to a Qt method accepting a
    WId parameter.

Since WId is only used for KDE::KWindowSystem methods here, I have not
used #ifdef Q_WS_X11 guards like in those two functions.

Signed-off-by: Rick W. Chen <stuffcorpse@archlinux.us>