Replace volume button with traditonal volume slider in Main Toolbar.
Following a suggestion from Nightrose and Sandsmark, I'm giving this a try.
It's basically our old volume slider, but shorter, and without the text label.
Aim is to improve usability, without hurting looks too much. Note: We should
probably use the same SVG for volume as we're using for progress.