Usability: Append video to playlist on single-click.
So, this had confused me already a number of times: The video applet
required double-click for adding the video to the playlist. Discoverablity
was bad with this, and it often made me think the applet was broken. Also,
I've changed AppendAndPlayImmediately to AppendAndPlay, which is the standard
behavior in Amarok.