Systray information now updates correctly for streams.
The problem was that we used engineNewMetadata(), and then took the metadata
from current track. At that point there is no guarantee that Meta itself
has already updated, as the observer call order is random. The right way is to use
Meta::Observer::metadataChanged(). NOTE: The same bug exists in other places.
BUG: 211748