Add custom labels to the changelog.
Custom labels support is now mostly implemented, and so I wanted to get it to trunk.
As it stands, you can read and write labels for sql tracks from the tag dialog.  Tags can be viewed in the playlist, but it's readonly at this point.  I plan on allowing inline editing there.

Labels are fetched from last.fm for Meta::File tracks currently, as a demonstration of how it works.  We need to discuss when to fetch from last.fm, and how to differentiate last.fm labels from sql labels.  This is on my todo list.

Other things I'm still planning on doing for 2.2.2:  Some sort of context view support, a better view for labels in the tag dialog, and perhaps querymaker support... though that scares me just a lot.