Improve dynamic playlist's behaviour on track change
Consider a dynamic playlist with 2 Previous tracks:

track 1
track 2
active track
track 4
track 5
track 6

Previously if you double click on anything after track 4,  every track in
between would be skipped (put behind the active track, as if they have
already been played). Now, the selected track would be moved to the active
position and the rest are kept in place. And similarly for tracks that
before the active position.