Fix bias widget "With" combobox display when loading saved biases
Problem: If you save a proportional bias matching an artist for example,
the value of the "With" field combobox is blank when that bias is reloaded.
It happens when Amarok is restarted, or change to Random bias and then back.

This is because the combobox is not populated upon its creation. Only after
it is populated the "With" field can be set to the saved setting.