Revert "Fix "Cancelling cover manager does not work""
This reverts commit 3e7cc7e5bc7002d252e90aada31f8ed02458371a.

My "fix" made the cover manager load covers very slowly, due to the
excessive event processing calls. Although you can now cancel the
operation, the slowness is just not worth it. I tried to make it work
both ways but it was a failure. I'll revert this for the moment and see
if a sleep or two can help me to figure out a satisfactory approach.