Make context applet colours resilient to different colour styles
This is related to the Wikipedia, Videoclip, and Lyrics applets as the
other applets don't (yet) use these methods.

There are 2 main colours we're dealing with here:
* the background colour of the main area of the applet
* the alternative background colour the applet might want (e.g. the table
  backgrounds of the Wikipedia applet, or when the Lyrics applet is in
  editing mode)

The "Magic (TM)" for determining the values and saturation of these
colours depending on whether the current style is dark or light has been
removed. That didn't work well anyway; instead it just goes back to the
basics. So if you still have problems with the contrast between
background and text it is the fault of the colour style. Every colour
style I've tried shows good contrast now.

BUG: 214930