Ensure adding new tracks does not mess up play history in random mode
This affects going back and forward between tracks in a playlist in
random mode while adding tracks. Previously if you had gone back to a
previous track, and then add some new tracks to the playlist, the track
you were at prior to the change will be lost in the randomised pool of
unplayed tracks. So going forward again a random track will be picked.