Fix for bug #210837. Fixes two problems with SqlAlbum::setCompilation(bool):
1. Duplicate album under compilations is another album
with the same title is already under compilations.

2. Removing an album from compilations can leave an empty album, if
another album the same artist already exists.

No. 1 was caused by the fact the a join key in mysql is useless if a
part of the joined key is null, because multiple entries can exist with
the other key(s). This is fixed by checking for an existing album and
moving the tracks to that album.

No. 2 was caused by the joined key on album with artist+name. the code
also assumed all the tracks belonged to the same artist. This is fixed
by moving the tracks to existing albums with the same title and artist
or creating a new album for each track-artist if more than one.