Show frames in the lyrics and video applets
This improves visual feedback on what is focused when you TAB/click around.
The alternative background is also altered to fit nicely inside the frame.

The idea for this is so that you can use the keyboard to focus on the
lyrics, and then start editing it; CTRL+S when you're done. This is just
an example (for the future).