podcasts: better plain text description handling, display subscribe-/pub-date
Now when converting plain text descriptions to html I replace \n by <br/>
and convert urls and mail addresses to proper links.

In the info widget I display the subscription date of a channel and the
publication date of an episode. I'm not 100% sure about the formatting
of all this information, though.