Fix lyrics applet height calculation when collapsing
o set the dummy item's width to be the same as the lyrics to account
  for possible scrollbar.
o set the dummy item's font to be the same as the lyrics.
o remove QFontMetrics as it does not deal with multiline / wrapping.
o content size should be extracted properly now (tested with KRuler)