Fix lyrics applet sometimes not clearing suggested urls when switching track
with cleanups and made code simpler; one proxy widget instead of two.

The suggested lyrics are not hidden when switching to another track with
cached lyrics. This may happen if you switch to some other applet (so
that the lyrics applet is not shown), whereby both m_lyrics and
m_suggested are hidden, and then switch back to the lyrics applet. The
previous logic used for determining which text browser is active breaks
down.