Try to work around a database leak in podcasts.
Because of a stupid mistake on my part there is bogus data in the podcastepisodes table. By ordering the results by desending id we should get the correct results first in the result.

This will solve the "limit episodes not honored" part but I think there will still be duplicate episode. The only fix for that issue is a database cleanup.
CCBUG:219516