Fix and future proof MusicBrainz track uid prefix
o Summary:

before "amarok-sqltrackuid://MB_75f0886c-ae10-432b-bdb9-c0e728f98621"
after  "amarok-sqltrackuid://mb-75f0886c-ae10-432b-bdb9-c0e728f98621"
                             ^^^

o Problem 1:

When a QString containing a uid gets converted to QUrl/KUrl, it is
automatically converted to lowercase. From Qt docs:

	Note that the case folding rules in Nameprep, which QUrl
	conforms to, require host names to always be converted to
	lower case, regardless of the Qt::FormattingOptions used.

The result is that whenever such a conversion occurs, the corresponding
track is no longer valid, causing it to be dropped. This happens, for
example, when loading the XSPF playlist and when the dynamic playlist
searches for a track. Dragging tracks with MBIDs from the album applet or
the collection browser works because it does not go through the same path
flow.

Consequently, the dynamic playlist will never load a track that was tagged
using MusicBrainz; a current.xspf with such tracks saved (could happen if
you drag them to the playlist and then close Amarok) will not show those
tracks when Amarok is reopened.

Instead of going through and manage these conversions wherever they pop up
I decided to change the prefix directly.

Addendum: it turns out the issue with dynamic playlists is plagued by
another bug. This is addressed in a separate patch.

o Problem 2 (Qt 4.6):

Due to changes to QUrl in 4.6, underscores are no longer allowed in the
hostname:

 - QUrl's parser is more strict when for hostnames in URLs. QUrl now
   enforces STD 3 rules:

    * each individual hostname section (between dots) must be at most
      63 ASCII characters in length;

    * only letters, digits, and the hyphen character are allowed in the
      ASCII range; letters outside the ASCII range follow the normal
      IDN rules

   That means QUrl no longer accepts some URLs that were invalid
   before, but weren't interpreted as such.

    http://qt.nokia.com/developer/changes/changes-4.6.0

CCBUG: 219902